Moved the code for each page to its own file
