+++ static/style.css
@@ -134,7 +134,6 @@ textarea {
     margin-left: 10px;
 }
 .date-section {
-    border-bottom: 1px solid #ddd;
     padding-bottom: 15px;
     margin-bottom: 25px;
 }
@@ -145,8 +144,76 @@ textarea {
     margin-bottom: 10px;
     background-color: #FFFFED;
     border: 1px solid #ccc;
+    padding: 5px;
 }
 
 .hidden {
     display: none;
+}
+
+/* To-Do List Styles */
+.todo-form-container {
+    margin-bottom: 20px;
+    padding: 15px;
+    border: 1px solid #ddd;
+    border-radius: 5px;
+    display: flex;
+    flex-wrap: wrap;
+    gap: 10px;
+}
+
+    .todo-form-container input,
+    .todo-form-container select,
+    .todo-form-container button {
+        padding: 8px;
+        border-radius: 3px;
+        border: 1px solid #ccc;
+    }
+
+.todo-list-container {
+    margin-top: 20px;
+}
+
+.project-section h3 {
+    background-color: #f0f0f0;
+    padding: 10px;
+    border-radius: 5px;
+}
+
+.todo-item {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 10px;
+    border-bottom: 1px solid #eee;
+}
+
+.item-content {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+}
+
+.priority {
+    font-size: 0.8em;
+    padding: 3px 6px;
+    border-radius: 3px;
+    color: white;
+}
+
+.priority-low {
+    background-color: #28a745;
+}
+
+.priority-medium {
+    background-color: #ffc107;
+}
+
+.priority-high {
+    background-color: #dc3545;
+}
+
+.status-finished .item-text {
+    text-decoration: line-through;
+    color: #888;
 }
