+++ templates/day_view.html
@@ -15,7 +15,7 @@
 <div class="entries-list-container">
     {% for entry in entries %}
     <div class="entry">
-        <a href="{{ url_for('edit', entry_id=entry.id) }}">
+        <a href="{{ url_for('calendar_bp.edit', entry_id=entry.id) }}">
             <span class="timestamp">{{ entry.timestamp.split(' ')[1] }}</span> - {{ entry.content }}
             {% if entry.time_elapsed %}
             <span class="time-elapsed">({{ entry.time_elapsed }})</span>
