+++ templates/index.html
@@ -16,7 +16,7 @@
         <ul class="entries-list" style="list-style-type: none; padding: 0;">
             {% for entry in entries %}
             <li class="entry">
-                <a href="{{ url_for('edit', entry_id=entry.id) }}">
+                <a href="{{ url_for('calendar_bp.edit', entry_id=entry.id) }}">
                     <span class="timestamp">{{ entry.timestamp.split(' ')[1] }}</span> - {{ entry.content }}
                     {% if entry.time_elapsed %}
                     <span class="time-elapsed">({{ entry.time_elapsed }})</span>
